organize traverser classes within Entity namespace
